Added protected variables in machine class.
Changed because we need to print both initial and current data in the final report text file. 
Added currentCoins, initialDollars, curretDollars, transactions, and totalCost variables and changed variables in OneItem. Changed related functions: constructor, setCoins, setItems, and printing functions. 
When we do the purchase, we can just change the "current" data, and leave initial data unchanged. 
And please only call setCoins and setItem function only during the initialization. 
The destructor is not finished, I am working on it...